# docs/swagger.yaml
...
host: localhost:8084
info:
  contact: {}
  description: FLIP backend server.
  termsOfService: http://swagger.io/terms/
  title: FLIP backend API
  version: "1.0"
paths:
...
  /v1/notoken/change/password:
    post:
      consumes:
      - text/plain
      description: change a user's password
      parameters:
      - description: email
        in: body
        name: email
        required: true
        schema:
          type: string
      - description: question
        in: body
        name: question
        required: true
        schema:
          type: integer
      - description: answer
        in: body
        name: answer
        required: true
        schema:
          type: string
      - description: newpwd
        in: body
        name: newpwd
        required: true
        schema:
          type: string
      - description: confirm
        in: body
        name: confirm
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"status":200, "message":update successfully}'
          schema:
            type: json
      summary: vertify a user's email and change password
...
securityDefinitions:
  JWT:
    type: basic
swagger: "2.0"